requirements.txt
pycryptodome == 3.23.0; python_version >= "3.7"